+++ src/translators/tests/conftest.py
@@ -3,7 +3,7 @@ import os
 from six.moves.urllib.error import HTTPError
 from six.moves.urllib.request import urlopen
 
-
+ 
 def check_crate(docker_ip, public_port):
     """Check if a crate is reachable.
 
@@ -34,6 +34,14 @@ def docker_stack(docker_services):
         4200,
         check_server=check_crate,
     )
+    # even though the http api is available, cratedb does not allow
+    # connections immediately. so sleep a little before creating a user
+    # to test crate authentication
+    from time import sleep
+    sleep(5)
+    docker_services.execute('crate', "bash", "-c",
+    "crash -c \"CREATE USER quantumleap WITH (password = 'a_secret_password');\" && \
+    crash -c \"GRANT DML,DDL,DQL TO quantumleap;\"")
 
 
 @pytest.fixture(scope='session')